+++ src/events/RegisterElementTypesEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace samuelreichoer\queryapi\events;
+
+use craft\base\Event;
+use samuelreichoer\queryapi\models\RegisterElementType;
+
+class RegisterElementTypesEvent extends Event
+{
+    /**
+     * @var RegisterElementType[]
+     */
+    public array $elementTypes = [];
+}